Move Configuration to new package
